pyproject.toml
[project]
name = "wsb-agent"
version = "0.1.0"
description = "Agentic system for analyzing WallStreetBets sentiment and generating trading signals"
requires-python = ">=3.11"
dependencies = [
    "praw>=7.7",
    "yfinance>=0.2",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "vaderSentiment>=3.3",
    "pandas>=2.0",
    "requests>=2.32.5",
    "alpaca-py>=0.43.2",
    "ollama>=0.6.1",
    "fastapi>=0.129.2",
    "uvicorn>=0.41.0",
]

[project.scripts]
wsb-agent = "wsb_agent.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/wsb_agent"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests requiring live API access",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
]
